Changed 'template' from a variable to a function that always returns a template filled in with content sent to it.
Change the default content to be entries of key/value pairs (labelled
key and value).

Render the controls in a more d3.way, by binding data to each control.
On exiting the control, the changed value is cached back into that data.

Wire up the save button to update the gist.